fixing tests on go1.9 and older
