+++ tds_test.go
@@ -176,16 +176,7 @@ func (l testLogger) Println(v ...interface{}) {
 	l.t.Log(v...)
 }
 
-func open(t *testing.T) *sql.DB {
-	checkConnStr(t)
-	SetLogger(testLogger{t})
-	connector, err := NewConnector(makeConnStr(t).String())
-	if err != nil {
-		t.Error("Open connection failed:", err.Error())
-		return nil
-	}
-	return sql.OpenDB(connector)
-}
+
 
 func TestConnect(t *testing.T) {
 	checkConnStr(t)
